Creating library testing verona-mlir
In the same way as `verona-ast` uses the AST library to test it, this
executable will use the MLIR library and test it. Right now, there is
only a skeleton implementation of the driver, no AST->MLIR testing.

Right now, the only thing it can do is read MLIR, validate (using the
Verona dialect) and spit it out again. But since the dialect is broken,
the output is (syntacticaly correct) garbage. If the input is in
standard dialects, it can also spit out LLVM IR. With tests.

Interesting features:
 * Standard cmd-line options (`-o`, `--emit-llvm`, etc)
 * Auto-detection of file suffix (`.mlir`, `.ll`, etc)
 * Reads MLIR, spits out MLIR and LLVM
 * CMake magic for single extra variable LLVM_BUILD

Work in progress:
 * Connected to the AST lib, but doesn't parse anything yet
 * Generator has a stub function to drive that through

Issues:
 * CMake files are a bit of a mess, mixing std and llvm stuff
 * Too many includes, need to do a clean up soon
 * Inconsistent error handling (cerr, throw)