Parsing Verona function names and signature
Finding function names from tokens, and argument and return types from
the signature, and lowering as an opaque "!type" type.

For now, we treat definitions as declarations, in order to test that
part, and there's no expression / basic block lowering at all.

The function.verona AST test was cloned to provide interim testing for
MLIR lowering.