LLVM submodule, build with artifact (Lin/Win/Mac)
Adding LLVM submodule on a specific commit and pulling the build
directory from DevOps so we don't have to build it on slow machines.

The scripts work on both Azure CI and desktop (Linux, Windows, Mac)

Option to disable LLVM for Format build. May need something more robust
in the future.